+++ src/test/java/kr/megaptera/assignment/applications/comments/CreatePostServiceTest.java
@@ -1,17 +1,12 @@
-package kr.megaptera.assignment.applications;
+package kr.megaptera.assignment.applications.comments;
 
 import jakarta.transaction.Transactional;
-import kr.megaptera.assignment.dtos.posts.PostCreateDto;
+import kr.megaptera.assignment.applications.posts.CreatePostService;
 import kr.megaptera.assignment.repositories.PostRepository;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import java.util.Optional;
-
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
 
 @SpringBootTest
