+++ src/test/java/kr/megaptera/assignment/applications/comments/GetPostsServiceTest.java
@@ -1,9 +1,9 @@
-package kr.megaptera.assignment.applications;
+package kr.megaptera.assignment.applications.comments;
 
+import kr.megaptera.assignment.applications.posts.GetPostsService;
 import kr.megaptera.assignment.dtos.posts.PostReadDto;
 import kr.megaptera.assignment.models.posts.MultilineText;
 import kr.megaptera.assignment.models.posts.Post;
-import kr.megaptera.assignment.models.posts.PostId;
 import kr.megaptera.assignment.repositories.PostRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
